Replace `rspec-rails` with `rspec`
https://github.com/thoughtbot/json_matchers/pull/39

This PR replaces `rspec-rails` with `rspec`, and removes the
`activesupport` dependency, which should make it easier to use this gem
outside of a Rails environment.

As a nice bonus, it also reduces the total number of gem dependencies
(as reported by `bundle install`) from 35 to 15.

There were only a few methods provided by ActiveSupport that were being
used, which have been replaced.