Add `factory_bot` for building test data
Define the `response` factory to replace the `response_for` helper.

To work around `FactoryBot`'s expectations for its arguments to be a
factory name, traits, then a Hash of attributes, the factory accepts a
`body:` for building invalid JSON or JSON that has an Array as its root.

Otherwise, the factory builds a dummy object that response to `#body`
that returns the attributes as JSON.