Declare `schema` factory
This commit declares a FactoryBot factory for creating a JSON schema.

Before the introduction of FactoryBot, the global `create_schema` test
helper was responsible for turning a name and a JSON hash into a JSON
Schema file, persisted to the filesystem.

All `create_schema` call sites required a full JSON schema as its second
argument.

Most `create_schema` call sites share duplicated JSON fragments, and
some covered similar or identical scenarios.

With the addition of factories, there is an opportunity to declare
[FactoryBot traits][traits] to abstract away the boring bits.

[traits]: https://github.com/thoughtbot/factory_bot/blob/master/GETTING_STARTED.md#traits