Introduce `referencing_objects` trait
The `"supports $ref"` test used to contain:

* expectations that the matcher validates valid JSON
* expectations that the matcher rejects invalid JSON
* expectations that both the `match_json_schema` and
  `match_response_schema` matchers were supported

Originally, the test was overloaded with assertions because constructing
a schema referencing another schema _and_ the referenced schema was
tedious and noisy.

With the adoption of using factories, this is no longer the case.

This commit splits out those concerns across multiple tests.