Rename factories to location
The `object` trait is too generic of a word to discuss unambiguously.

For example, JSON Schema supports `{ "type": "object" }` declarations,
and JSON "hashes" are "objects" as well.

The README currently uses latitude and longitude as an example.

This commit modifies the factories to build latitude-longitude pairs
instead of objects with `id` values.

Co-authored-by: Sean Doyle <sean.p.doyle24@gmail.com>
Co-authored-by: Daniel Colson <danieljamescolson@gmail.com>