Add coverage for JSON Schema V4
Closes [#88].
Closes [#90].

Add tests from Issue #90

When the `"id"` key is specified as a `"number"` in both attribute and
property definitions, then specified as a `"string"` value in the
payload, the matcher will raise and the test will pass.

After rebasing off `master` after [`3b66b4d`][3b66b4d] was merged, [the
tests that were once failing][tests] are now passing.

[#88]: https://github.com/thoughtbot/json_matchers/issues/88
[#90]: https://github.com/thoughtbot/json_matchers/issues/90
[3b66b4d]: https://github.com/thoughtbot/json_matchers/commit/3b66b4ddf369ec09b50cfe39c614266b34c7f3fe
[tests]: https://github.com/thoughtbot/json_matchers/pull/89#issuecomment-437484000